+++ gammapy/estimators/__init__.py
@@ -1,7 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Estimators."""
 from .parameter_estimator import *
-from .lima_map import *
+from .excess_map import *
 from .ts_map import *
 from .asmooth_map import *
 from .sensitivity import *
