+++ gammapy/estimators/excess_map.py
