+++ gammapy/estimators/tests/test_excess_map.py
